Refactoring: removed unnecessary class and imports
* Removed unnecessary imports
* Fixed PMD and checkstyle issues
* merged controller/services/EntryService.java into entry/EntryService.java
